+++ osmnx/_api.py
@@ -28,6 +28,13 @@ from .graph import graph_from_file
 from .graph import graph_from_place
 from .graph import graph_from_point
 from .graph import graph_from_polygon
+from .graph import graph_from_xml
+
+from .io import load_graphml
+from .io import save_graph_geopackage
+from .io import save_graph_shapefile
+from .io import save_graph_xml
+from .io import save_graphml
 
 from .plot import plot_figure_ground
 from .plot import plot_graph
@@ -41,12 +48,6 @@ from .pois import pois_from_point
 from .pois import pois_from_polygon
 
 from .projection import project_graph
-
-from .save_load import load_graphml
-from .save_load import save_graph_geopackage
-from .save_load import save_graph_osm
-from .save_load import save_graph_shapefile
-from .save_load import save_graphml
 
 from .simplification import clean_intersections
 from .simplification import consolidate_intersections
@@ -64,4 +65,5 @@ from .utils_geo import geocode
 
 from .utils_graph import gdfs_to_graph
 from .utils_graph import get_undirected
+from .utils_graph import graph_from_gdfs
 from .utils_graph import graph_to_gdfs
